+ Move proto to independent folder.
