+++ packages/core/src/parser/svga.ts
@@ -1,13 +1,12 @@
 import pako from 'pako'
-import { IProtoMovieEntity } from '../proto/models'
-import { decodeMovieEntity } from '../proto/svga'
+import {decodeMovieEntity} from '@svgaplus/proto'
+import {IProtoMovieEntity} from '@svgaplus/proto/types'
 
 abstract class SVGAParser {
   static async parse (svgaBuffer: ArrayBuffer): Promise<IProtoMovieEntity> {
     const bufferArray = new Uint8Array(svgaBuffer)
     const buffer = pako.inflate(bufferArray)
-    const movieEntity = decodeMovieEntity(buffer)
-    return movieEntity
+    return decodeMovieEntity(buffer)
   }
 }
