+++ packages/proto/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@svgaplus/proto",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Protobuf of SVGA",
+  "scripts": {
+    "pb": "npx pbf ./svga.proto > ./svga.js",
+    "test": "jest"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.2.1",
+    "jest": "^29.2.2",
+    "pbf": "^3.2.1",
+    "ts-jest": "^29.0.3",
+    "typescript": "^4.8.4"
+  }
+}
