+++ packages/proto/svga.js
