+++ packages/proto/tsconfig.json
@@ -19,6 +19,5 @@
       "es2017",
       "es2018"
     ]
-  },
-  "exclude": ["test"]
+  }
 }
