+++ packages/renderer.pixi/tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
     "moduleResolution": "node",
@@ -15,10 +16,6 @@
       "es2016",
       "es2017",
       "es2018"
-    ],
-    "paths": {
-      "@svgaplus/core": ["../core/src"],
-      "@svgaplus/core/types/*": ["../core/types/*"],
-    }
+    ]
   }
 }
