+++ showcase/pages/upload/index.html
@@ -3,6 +3,7 @@
     <meta charset="utf-8">
     <style>
       canvas {
+        max-width: 650px;
         background-color: #000;
       }
     </style>
