+++ showcase/vite.config.ts
@@ -5,6 +5,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@svgaplus/core': path.resolve(__dirname, '../packages/core/src/index.ts'),
+      '@svgaplus/proto': path.resolve(__dirname, '../packages/proto'),
       '@svgaplus/renderer.pixi': path.resolve(__dirname, '../packages/renderer.pixi/lib/index.ts')
     }
   }
